Moved repositories to their own package Added test data Added TODO in views
Application is not runnable after debugging

+++ src/main/java/com/example/application/data/Moon.java
@@ -10,7 +10,7 @@ public class Moon extends AbstractEntity {
     private String moonName;
 
     @NotNull
-    private int cycle;
+    private double cycle;
 
     @NotNull
     private int shift;
@@ -29,11 +29,11 @@ public class Moon extends AbstractEntity {
         this.moonName = moonName;
     }
 
-    public int getCycle() {
+    public double getCycle() {
         return cycle;
     }
 
-    public void setCycle(int cycle) {
+    public void setCycle(double cycle) {
         this.cycle = cycle;
     }
